wifi: mt76: mt7915: do not use event format to get survey data
Using the event format to get survey data results in the chip
eventually crashing with no chance of recovery. The only possible
course of action is to restart the system and hope it never happens
again.

Also the event format doesn't function properly with OEM WM fw,
making it impossible to use OEM WM fw with the open-source driver.

Signed-off-by: rany <ranygh@riseup.net>